New API, builders and classes structure.
Trees are now provided in two main flavours: undirected and directed.
The mathematically compliant tree implementation is
RawUndirectedTreeBuilder, which is available as a class under the
Tree name. Another useful implementation is the
RawDirectedTreeBuilder, which is available as a class under the
Arborescence name.

Beware this breaks previous API and namespacing.